requirements.txt
agent-framework==1.0.0b251001
agent-framework-azure-ai==1.0.0b251001
python-dotenv>=1.0.1
pydantic>=2.8.2
tenacity>=9.0.0
azure-identity>=1.17.1
ruff>=0.6.9
black>=24.8.0
mypy>=1.10.0
pre-commit>=3.8.0
